Check remote golang cli check
