+++ examples/basic/variables.tf
@@ -37,7 +37,3 @@ variable "resource_tags" {
   description = "Optional list of tags to be added to created resources"
   default     = []
 }
-
-####################
-# Optional variables
-####################
