--- scripts/confirm-rollout-status.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-set -e
-
-daemonset=$1
-namespace=$2
-
-kubectl rollout status ds "${daemonset}" -n "${namespace}" --timeout 30m
